Split into two packages to allow importing as library
