+++ Makefile
@@ -11,7 +11,7 @@ endif
 
 NAME:=$(notdir $(CURDIR))
 VERSION:=$(shell git describe --tags 2>$(NUL) || echo v0.0.0)
-GOOPT:=-ldflags "-s -w -X main.version=$(VERSION)"
+GOOPT:=-ldflags "-s -w -X core.version=$(VERSION)"
 EXE:=$(shell go env GOEXE)
 
 all:
